Split up nbinspect into formats
