+++ appyter/render/nbinspect/__init__.py
@@ -1,41 +1,15 @@
 import os
 import json
 import click
-import nbformat as nbf
 
 from appyter.cli import cli
-from appyter.util import dict_filter_none
 from appyter.ext.fs import Filesystem
 from appyter.parse.nb import nb_from_ipynb_io
-from appyter.parse.nbtemplate import parse_fields_from_nbtemplate
 from appyter.context import get_env, get_jinja2_env
 from appyter.ext.click import click_option_setenv, click_argument_setenv
 
-def render_nbtemplate_json_from_nbtemplate(env, nb):
-  ''' Render a json representing the relevant Fields throughout the notebook.
-  '''
-  return [
-    { 'field': field.field, 'args': field.args }
-    for field in parse_fields_from_nbtemplate(env, nb, deep=True)
-  ]
-
-def render_jsonschema_from_nbtemplate(env, nb):
-  ''' Render a jsonschema representing the relevant Fields throughout the notebook.
-  TODO: branch with tabs
-  '''
-  fields = list(parse_fields_from_nbtemplate(env, nb, deep=True))
-  return {
-    'type': 'object',
-    'properties': dict_filter_none({
-      field.args['name']: field.to_jsonschema()
-      for field in fields
-    }),
-    'required': [
-      field.args['name']
-      for field in fields
-      if field.args.get('required')
-    ],
-  }
+from appyter.render.nbinspect.nbtemplate_json import render_nbtemplate_json_from_nbtemplate
+from appyter.render.nbinspect.jsonschema import render_jsonschema_from_nbtemplate
 
 @cli.command(help='Inspect appyter for arguments (fields)')
 @click.option('-o', '--output', default='-', type=click.File('w'), help='The output location of the inspection json')
